refactor(UIKIT-1275): Исправления по рефакторингу
