--- packages/components/src/FormHelperText/Content/index.ts
@@ -1,1 +1,0 @@
-export * from './Content';
