+++ packages/components/src/FormHelperText/FormHelperTextContent/index.ts
@@ -1,3 +1,1 @@
-export * from './FormHelperText';
-
 export * from './FormHelperTextContent';
