+++ packages/components/src/NotificationList/Empty/index.ts
@@ -1,0 +1,1 @@
+export * from './Empty';
