+++ packages/components/src/NotificationList/ListItem/index.ts
@@ -1,0 +1,1 @@
+export * from './ListItem';
