--- packages/components/src/NotificationList/NotificationListEmpty/index.ts
@@ -1,1 +1,0 @@
-export * from './NotificationListEmpty';
