--- packages/components/src/NotificationList/NotificationListItem/index.ts
@@ -1,1 +1,0 @@
-export * from './NotificationListItem';
