--- packages/components/src/NotificationList/NotificationListSettingsButton/index.ts
@@ -1,1 +1,0 @@
-export * from './NotificationListSettingsButton';
