+++ packages/components/src/NotificationList/SettingsButton/index.ts
@@ -1,0 +1,1 @@
+export * from './SettingsButton';
